+++ flags.nix
@@ -1,0 +1,9 @@
+{
+  "exceptions" = true;
+  "logict" = true;
+  "monad-logger" = true;
+  "mtl" = true;
+  "primitive" = true;
+  "resourcet" = true;
+  "unliftio" = true;
+}